requirements.txt
# ecg_vqa_system/requirements.txt
torch>=2.0.0
transformers>=4.30.0
efficientnet-pytorch==0.7.1
pillow>=9.0.0
tqdm>=4.65.0
scikit-learn>=1.2.0